% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_target_ownship_truth.R
\name{combine_target_ownship_truth}
\alias{combine_target_ownship_truth}
\title{Returns dataframe with target and ownship truth combined}
\usage{
combine_target_ownship_truth(scenario)
}
\arguments{
\item{scenario}{scenario}
}
\value{
a position dataframe, which will include:
\itemize{
\item{time}
\item{lat}
\item{lon}
\item{alt}
\item{heading}
\item{truthID}
\item{extra columns either dataframe has}
}
}
\description{
Creates Returns dataframe with target and ownship truth combined - is okay if one of the dataframes is NA
}
