% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example1_engagementData}
\alias{example1_engagementData}
\title{Engagement data for example 1}
\format{A data frame with 13 observations and 6 variables:
\describe{
  \item{time}{Time of weapon fire}
  \item{source}{Source originiating the shot}
  \item{target}{Target the shot was aimed at}
  \item{weapon}{Code for which weapon was used}
  \item{color}{Color in which to plot the shot}
  \item{kill}{Flag for whether target was killed by shot (0 or 1)}

}}
\usage{
example1_engagementData
}
\description{
Engagement data for example 1.  Used by SIMDIS and sandTable
}
\keyword{datasets}
