% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example1_ownShipData}
\alias{example1_ownShipData}
\title{Ownship data for example 1}
\format{A data frame with 526 observations and 6 variables:
\describe{
  \item{lat}{Latitude (degrees)}
  \item{lon}{Longitude (degrees)}
  \item{time}{Time of position measurement (double)}
  \item{alt}{Altitude (meters)}
  \item{heading}{Ownship heading (degrees azimuth)}
  \item{truthID}{Ownship ID is just "USS Good Guy" in every row (factor)}

}}
\usage{
example1_ownShipData
}
\description{
Example 1 has 8 targets and 14 tracks.  Targets move at realistic speeds.  Some are destroyed partway through the test
}
\keyword{datasets}
