% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example1_platformInfo}
\alias{example1_platformInfo}
\title{Platform info for example 1}
\format{
A data frame with 23 observations and 4 variables:
\describe{
  \item{truthID}{Target ID for each measurement.  Factor.  Values are "A" through "H"}
  \item{platformIcon}{Icon used for each platform (Used by SIMDIS)}
  \item{platformType}{Type of platform (used by SIMDIS)}
  \item{trackColor}{Color of track}

}
}
\usage{
example1_platformInfo
}
\description{
Platform info for example 1.  Used by SIMDIS and sandTable
}
\keyword{datasets}
