% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example1_scenario}
\alias{example1_scenario}
\title{Example 1 scenario}
\format{
A named list of 9 elements containing:
\describe{
  \item{scenarioName}{Scenario name (string)}
  \item{sensorData}{see help page for example1_sensorData for more details}
  \item{targetTruth}{see help page for example1_truthData for more details}
  \item{ownShipTruth}{see help page for example1_ownShipData for more details}
  \item{engagementData}{see help page for example1_engagementData for more details}
  \item{platformInfo}{see help page for example1_platformInfo for more details}
  \item{targetTrackDistance}{Calculated from targetTruth, sensorData, and ownShipTruth using the function target_track_distance().  See help page for target_track_distance() for more details}
  \item{targetOwnShipDistance}{Calculated from targetTruth and ownShipTruth using the function target_ownship_distance().  See help page for target_ownship_distance() for more details}
  \item{trackOwnShipDistance}{Calculated from sensorData and ownShipTruth using the function track_ownship_distance().  See help page for track_ownship_distance() for more details}

}
}
\usage{
example1_scenario
}
\description{
Example 1 scenario.  Example 1 has 8 targets and 14 tracks.  Targets move at realistic speeds.  Some are destroyed partway through the test
}
\keyword{datasets}
