% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example1_sensorData}
\alias{example1_sensorData}
\title{Sensor data for example 1}
\format{
A data frame with 2312 observations and 5 variables:
\describe{
  \item{lat}{Latitude (degrees)}
  \item{lon}{Longitude (degrees)}
  \item{time}{Time of position measurement (double)}
  \item{alt}{Altitude (meters)}
  \item{trackNum}{Track ID (factor)}

}
}
\usage{
example1_sensorData
}
\description{
Example 1 has 8 targets and 14 tracks.  Targets move at realistic speeds.  Some are destroyed partway through the test
}
\keyword{datasets}
