% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example1_truthData}
\alias{example1_truthData}
\title{Truth data for example 1}
\format{
A data frame with 5010 observations and 6 variables:
\describe{
  \item{lat}{Latitude (degrees)}
  \item{lon}{Longitude (degrees)}
  \item{time}{Time of position measurement (double)}
  \item{alt}{Altitude (meters)}
  \item{heading}{Target heading (degrees azimuth)}
  \item{truthID}{Target ID for each measurement.  Factor.  Values are "A" through "H"}

}
}
\usage{
example1_truthData
}
\description{
Example 1 has 8 targets and 14 tracks.  Targets move at realistic speeds.  Some are destroyed partway through the test
}
\keyword{datasets}
