% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example2_ownShipData}
\alias{example2_ownShipData}
\title{Ownship data for example 2}
\format{
A data frame with 300 observations and 6 variables:
\describe{
  \item{lat}{Latitude (degrees)}
  \item{lon}{Longitude (degrees)}
  \item{time}{Time of position measurement (double)}
  \item{alt}{Altitude (meters)}
  \item{heading}{Ownship heading (degrees azimuth)}
  \item{truthID}{Ownship ID is just "USS Good Guy" in every row (factor)}

}
}
\usage{
example2_ownShipData
}
\description{
Example 2 has 2 targets and 1 track.
}
\keyword{datasets}
