% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example2_scenario}
\alias{example2_scenario}
\title{Example 2 scenario}
\format{A named list of 9 elements containing:
\describe{
  \item{scenarioName}{Scenario name (string)}
  \item{sensorData}{see help page for example2_sensorData for more details}
  \item{targetTruth}{see help page for example2_truthData for more details}
  \item{ownShipTruth}{see help page for example2_ownShipData for more details}
  \item{engagementData}{NA - This was not an input to the original call to create_scenario()}
  \item{platformInfo}{NA - This was not an input to the original call to create_scenario()}
  \item{targetTrackDistance}{Calculated from targetTruth, sensorData, and ownShipTruth using the function target_track_distance().  See help page for target_track_distance() for more details}
  \item{targetOwnShipDistance}{Calculated from targetTruth and ownShipTruth using the function target_ownship_distance().  See help page for target_ownship_distance() for more details}
  \item{trackOwnShipDistance}{Calculated from sensorData and ownShipTruth using the function track_ownship_distance().  See help page for track_ownship_distance() for more details}

}}
\usage{
example2_scenario
}
\description{
Example 2 scenario.  Example 2 has 2 targets and 1 track.  Targets move at realistic speeds.
}
\keyword{datasets}
