% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example2_sensorData}
\alias{example2_sensorData}
\title{Sensor data for example 2}
\format{
A data frame with 21 observations and 6 variables:
\describe{
  \item{lat}{Latitude (degrees)}
  \item{lon}{Longitude (degrees)}
  \item{time}{Time of position measurement (double)}
  \item{alt}{Altitude (meters)}
  \item{trackNum}{Track ID (factor)}
  \item{heading}{Ownship heading (degrees azimuth)}

}
}
\usage{
example2_sensorData
}
\description{
Example 2 has 2 targets and 1 track.
}
\keyword{datasets}
