% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example2_truthData}
\alias{example2_truthData}
\title{Truth data for example 2}
\format{A data frame with 337 observations and 6 variables:
\describe{
  \item{lat}{Latitude (degrees)}
  \item{lon}{Longitude (degrees)}
  \item{time}{Time of position measurement (double)}
  \item{alt}{Altitude (meters)}
  \item{heading}{Target heading (degrees azimuth)}
  \item{truthID}{Target ID for each measurement.  Factor.  Values are "A" through "H"}

}}
\usage{
example2_truthData
}
\description{
Example 2 has 2 targets and 1 track.
}
\keyword{datasets}
