% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_scenario_to_csv.R
\name{export_scenario_to_csv}
\alias{export_scenario_to_csv}
\title{Export contents of scenario to csv}
\usage{
export_scenario_to_csv(scenario, saveLocation, saveName = NULL)
}
\arguments{
\item{scenario}{will save the following dataframes to disk (ignoring those not present in the scenario): targetTruth, ownShipTruth, sensorData, engagementData, platformInfo, targetOwnShipDistance}

\item{saveLocation}{string - path to save file}

\item{saveName}{string (default=NULL) if no argument passed, will use the scenario's scenarioName property}
}
\description{
Export contents of scenario to csv.  Each data frame in the list is saved as a separate .csv
}
\examples{
\dontrun{export_scenario_to_csv(example1_scenario,saveLocation="C:/Documents/analysis/", saveName="testPeriod1")}
}
