% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesic_lat_long_offset.R
\name{geodesic_lat_long_offset}
\alias{geodesic_lat_long_offset}
\title{Geodesic lat lon offset}
\usage{
geodesic_lat_long_offset(referenceLat, referenceLon, offsetLat, offsetLon)
}
\arguments{
\item{referenceLat}{reference latitude}

\item{referenceLon}{reference longitude}

\item{offsetLat}{offset latitude}

\item{offsetLon}{offset longitude}
}
\value{
A geodesic line (see geosphere package) between the reference point and the offset point calculated on a spheroidal earth (WGS84) by geosphere package.  This function is used by transform_offset_to_latlon() and is not recommended for general use
}
\description{
This function is used by transform_offset_to_latlon() and it is not recommended that the user try to use this function as a stand-alone function.
}
\keyword{internal}
