% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dups_between_groups.R
\name{get_dups_between_groups}
\alias{get_dups_between_groups}
\title{Find duplicates between groups in a dataframe}
\usage{
get_dups_between_groups(df, idcol)
}
\arguments{
\item{df}{the dataframe you want to examine}

\item{idcol}{the column which identifies the group each row belongs to}
}
\value{
Return the vector of which entries are duplicated across groups
}
\description{
Identify identical rows in a dataframe belonging to different groups. Used by Nautilus's plot_compare_track_status
}
