% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_relative_distance.R
\name{get_relative_distance}
\alias{get_relative_distance}
\title{Get distance info between specified targets at the times provided}
\usage{
get_relative_distance(refPlatTruth, targetTruth)
}
\arguments{
\item{refPlatTruth}{data frame containing all of the truth position of the reference platform (ownShip).  MUST have the following columns:
\itemize{
 \item{time: (double) time of measurement. We currently recommend POSIX}
 \item{lon: (double) longitude of target at time of measurement}
 \item{lat: (double) latitude of target at time of measurement}
 \item{alt: (double) altitude of target at time of measurement}
 \item{truthID: (factor) name or identifier for target. We recommend letters or names.  This must contain ONE AND ONLY ONE VALUE FOR THIS FUNCTION}
 \item{heading: (double) sensor system heading in degrees azimuth}
}}

\item{targetTruth}{truthData: data frame containing all of the truth data for each target. MUST have the following columns:
\itemize{
 \item{time: (double) time of measurement. We currently recommend POSIX}
 \item{lon: (double) longitude of target at time of measurement}
 \item{lat: (double) latitude of target at time of measurement}
 \item{alt: (double) altitude of target at time of measurement}
 \item{truthID: (factor) name or identifier for target. We recommend letters or names.  This must contain ONE AND ONLY ONE VALUE FOR THIS FUNCTION}
 \item{heading: (double) target heading in degrees azimuth}
 }}
}
\value{
A dataframe containing the ranges and bearings from the sensor system to every target at each time the position of the sensor system was measured. The output columns are a limited version of target_track_distance() and includes:
\itemize{
  \item{time: time of the measurement of ownship position}
  \item{ownShipTruthID: unique identifier for the reference platform}
  \item{ownShipLon: longitude of ownship at this time}
  \item{ownShipLat: latitude of ownship at this time}
  \item{ownShipAlt: altitude of ownship at this time}
  \item{slantRange: range from ownship to the target (straight-line distance)}
  \item{groundRange: range from ownship to the target (everyone's altitude is zeroed out)}
  \item{targetAspect: the target aspect (as seen from ownship) at this time}
  \item{trueBearingToTarget: true bearing (azimuth) to the target at this time}
  \item{relBearingToTarget: relative bearing (azimuth) to the target at this time}
  \item{targetTruthID: unique identifier for the target}
  \item{targetLon: target longitude at this time}
  \item{targetLat: target latitude at this time}
  \item{targetAlt: target altitude at this time}
  \item{targetHeading: target heading at this time}
  }
}
\description{
Get distance info between specified targets at the times provided. Designed to work with exactly one truthID in each input dataframe. Both input dataframes MUST be on the same time base.  Be careful if you decide to use this; it is mostly intended as an internal function for distance_between() and target_ownship_distance().  If you're looking at this function, it's likely that you actually want one of those.
}
