% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_heading.R
\name{interp_heading}
\alias{interp_heading}
\title{Interpolate heading}
\usage{
interp_heading(currentTimes, currentHeadings, newTimes)
}
\arguments{
\item{currentTimes}{vector of the times of the existing heading measurements}

\item{currentHeadings}{vector of the existing heading measurements}

\item{newTimes}{vector of the times to be interpolated to}
}
\value{
a vector of headings at the new times (newTimes)
}
\description{
Function to interpolate headings.  Heading interpolation can be a little tricky when you wrap around from 0 to 360, so this function takes care of that.

Used by distance_between(), target_track_distance(), target_ownship_distance(), and track_ownship_distance().  It's likely that if you're looking at using this function you actually want one of those.
}
