% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance_data.R
\name{plot_distance_data}
\alias{plot_distance_data}
\title{Plot distance data}
\usage{
plot_distance_data(scenario, hideLegend = FALSE, useDefaultColors = FALSE)
}
\arguments{
\item{scenario}{must contain the output of target_ownship_distance() as a named object in the scenario list with the name "targetOwnShipDistance".  If you originally ran create_scenario() with both own ship data and target truth data then this should already be in the scenario.}

\item{hideLegend}{hides the legend.  May make the graph easier to read in scenarios with a lot of targets (default=FALSE)}

\item{useDefaultColors}{(default=FALSE) if you DO have platformInfo filled in, but want to ignore the colors and use defaults, set this to true.  If you do not have platformInfo, it will use default colors no matter what.}
}
\value{
ggplot object
}
\description{
This figure plots target range as a function of time
}
\examples{
plot_distance_data(example1_scenario, hideLegend = FALSE, useDefaultColors = FALSE)

}
