% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance_data_plotly.R
\name{plot_distance_data_plotly}
\alias{plot_distance_data_plotly}
\title{Plot distance data with plotly}
\usage{
plot_distance_data_plotly(
  scenario,
  useDefaultColors = FALSE,
  legendOnly = FALSE
)
}
\arguments{
\item{scenario}{must contain the output of target_ownship_distance() as a named object in the scenario list with the name "targetOwnShipDistance".  If you originally ran create_scenario() with both own ship data and target truth data then this should already be in the scenario.}

\item{useDefaultColors}{(default=FALSE) if you DO have platformInfo filled in, but want to ignore the colors and use defaults, set this to true}

\item{legendOnly}{(boolean, default = FALSE) Initially populates plot with only the legend.  Data can be turned on by selecting the data from the legend.  Good for plots with a large number of traces}
}
\value{
plotly object
}
\description{
This figure plots target range as a function of time.  Plotly figures are interactive, and users can zoom in, hover over data points for more info, and click to hide tracks or targets.
}
\examples{

plot_distance_data_plotly(example1_scenario, useDefaultColors = FALSE)
}
