% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relative_truth_and_sensor.R
\name{plot_relative_truth_and_sensor}
\alias{plot_relative_truth_and_sensor}
\title{Plot truth and/or sensor data relative to ownship}
\usage{
plot_relative_truth_and_sensor(
  scenario,
  offset = 1,
  textSize = 3,
  hideLegend = FALSE,
  useDefaultColors = FALSE,
  plotWhich = "both"
)
}
\arguments{
\item{scenario}{must contain either or both: targetOwnshipDistance, trackOwnshipDistance}

\item{offset}{(default = 1) amount to offset labels (in degrees)}

\item{textSize}{(default = 3) size of label text}

\item{hideLegend}{hides the legend to make the graph easier to read (default=FALSE)}

\item{useDefaultColors}{(default=FALSE) if you DO have platformInfo filled in, but want to ignore the colors and use defaults, set this to true}

\item{plotWhich}{(default = 'both') also accepts 'sensor' or 'truth'.  Plots only truth-ownship or track-ownship or plots both.}
}
\value{
ggplot object
}
\description{
This figure shows the bearing and range of targets and/or tracks relative to the sensor system (ownship).
}
\examples{
plot_relative_truth_and_sensor(example1_scenario, offset = 1, textSize = 3)
}
