% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relative_truth_and_sensor_plotly.R
\name{plot_relative_truth_and_sensor_plotly}
\alias{plot_relative_truth_and_sensor_plotly}
\title{Plot truth and/or sensor data relative to ownship - plotly version}
\usage{
plot_relative_truth_and_sensor_plotly(scenario, offset = 1, textSize = 3,
  useDefaultColors = FALSE, plotWhich = "both", legendOnly = FALSE)
}
\arguments{
\item{scenario}{must contain either or both: targetOwnshipDistance, trackOwnshipDistance}

\item{offset}{(default = 1) amount to offset labels (in degrees)}

\item{textSize}{(default = 3) size of label text}

\item{useDefaultColors}{(default=FALSE) if you DO have platformInfo filled in, but want to ignore the colors and use defaults, set this to true}

\item{plotWhich}{(default = 'both') also accepts 'sensor' or 'truth'.  Plots only truth-ownship or track-ownship or plots both.}

\item{legendOnly}{(boolean, default = FALSE) Initially populates plot with only the legend.  Data can be turned on by selecting the data from the legend.  Good for plots with a large number of traces}
}
\value{
plotly object
}
\description{
This figure shows the bearing and range of targets and/or tracks relative to the sensor system (ownship).
}
\examples{
plot_relative_truth_and_sensor_plotly(example1_scenario, offset = 1, textSize = 3)
}
