% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sensor_and_truth_data.R
\name{plot_sensor_and_truth_data}
\alias{plot_sensor_and_truth_data}
\title{Plot both sensor and truth data}
\usage{
plot_sensor_and_truth_data(scenario, offset = 0.01, textSize = 3,
  hideLegend = FALSE, useDefaultColors = FALSE)
}
\arguments{
\item{scenario}{MUST contain one of: targetTruth, ownShipTruth, and sensorData}

\item{offset}{(default = 0.01) amount to offset labels (in degrees latitude)}

\item{textSize}{(default = 3) size of label text}

\item{hideLegend}{hides the legend to make the graph easier to read (default=FALSE)}

\item{useDefaultColors}{(default=FALSE) if you DO have platformInfo filled in, but want to ignore the colors and use defaults, set this to true}
}
\value{
ggplot object
}
\description{
This figure shows an overview of the test and plots the ownship data (thin black line) along with the truth data (solid lines) and the tracks from the sensor (points with connecting lines). The target labels (likely characters or names) are placed above the the first point in time by an amount equal to the offset (in degrees latitude). The track labels (likely numbers) are placed below the last point in time of each track by the same amount. The size of the label text is given by textSize.

This figure can be created before any tracks are assigned to targets to ensure that the user has imported the data correctly and to get an overall picture of the test.
}
\examples{
plot_sensor_and_truth_data(example1_scenario, offset = 0.01, textSize = 3)
}
