% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sensor_and_truth_data_with_altitude.R
\name{plot_sensor_and_truth_data_with_altitude}
\alias{plot_sensor_and_truth_data_with_altitude}
\title{Plot both sensor and truth data with altitude}
\usage{
plot_sensor_and_truth_data_with_altitude(scenario, offset = 0.01,
  textSize = 3)
}
\arguments{
\item{scenario}{MUST contain targetTruth, ownShipTruth, and sensorData}

\item{offset}{amount to offset labels (in degrees latitude)}

\item{textSize}{size of label text}
}
\value{
layout of ggplots
}
\description{
In data sets with nonzero altitude it may be useful to plot target altitude along with latitude and longitude. The central panel of this figure is the same as plot_raw_data(), but includes side panels that show the altitude of tracks and targets.

This figure can be created before any tracks are assigned to targets to ensure that the user has imported the data correctly and to get an overall picture of the test.  Because it is a layout of objects it is more difficult to change the text labels than plot_sensor_and_truth_data().
}
\examples{
plot_sensor_and_truth_data_with_altitude(example1_scenario, offset = 0.01, textSize = 3)
}
