% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_truth_data_plotly.R
\name{plot_truth_data_plotly}
\alias{plot_truth_data_plotly}
\title{Plot truth data with plotly}
\usage{
plot_truth_data_plotly(scenario, useDefaultColors = FALSE,
  legendOnly = FALSE)
}
\arguments{
\item{scenario}{must contain either or both: targetTruth, ownShipTruth}

\item{useDefaultColors}{(default=FALSE) if you DO have platformInfo filled in, but want to ignore the colors and use defaults, set this to true}

\item{legendOnly}{(boolean, default = FALSE) Initially populates plot with only the legend.  Data can be turned on by selecting the data from the legend.  Good for plots with a large number of traces}
}
\value{
plotly object
}
\description{
This function shows an overview of the test and plots the ownship data and/or the truth data in a plotly figure.  Plotly figures are interactive, and users can zoom in, hover over data points for more info, and click to hide tracks or targets.
}
\examples{
plot_truth_data_plotly(scenarioMaker::example1_scenario)
}
