% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_truth_gaps.R
\name{plot_truth_gaps}
\alias{plot_truth_gaps}
\title{Plot the gaps in truth data}
\usage{
plot_truth_gaps(scenario)
}
\arguments{
\item{scenario}{must contain either or both: targetTruth, ownShipTruth}
}
\value{
ggplot object
}
\description{
This figure plots the gaps in truth data by target. Takes 1 or 2 dataframes, intended to plot both truth and/or ownship data
}
\examples{
plot_truth_gaps(example1_scenario)
}
