% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_time_exists.R
\name{summarize_time_exists}
\alias{summarize_time_exists}
\title{Summarize truth data for targets}
\usage{
summarize_time_exists(scenario)
}
\arguments{
\item{scenario}{must contain targetTruth and/or ownShipTruth}
}
\value{
data frame containing a summary of how long each target existed (i.e., how much truth data exists for each target)

\itemize{
 \item{"truthID": Name of target}
 \item{"startTime": First recorded time for each target}
 \item{"stopTime": Last recorded time for each target}
 \item{"timeExists": difference between stopTime and startTime (in seconds)}}
}
\description{
Table showing how long each target existed (according to truth data)
}
\examples{
summarize_time_exists(example1_scenario)
}
