% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_truth_gaps.R
\name{summarize_truth_gaps}
\alias{summarize_truth_gaps}
\title{Summarize the gaps in truth/ownship data}
\usage{
summarize_truth_gaps(scenario)
}
\arguments{
\item{scenario}{must contain either or both: targetTruth, ownShipTruth, which MUST have the following columns:
\itemize{
 \item{time: (double) time of measurement. We currently recommend POSIX}
 \item{truthID: (factor) name or identifier for target. We recommend letters or names}
}}
}
\value{
Data frame containing a summary of how much time each target was tracked.  The data frame contains the collowing variables:

\itemize{
 \item{"tgt" - Target name/identifier}
 \item{"# Intervals Above 2x Median" - The number of track update intervals longer than twice the median}
 \item{"# Intervals Above 5x Median" - The number of track update intervals longer than 5 times the median}
 \item{"# Intervals Above 10x Median" - The number of track update intervals longer than 10 times the median}
 \item{"Total # Intervals" - The total number of intervals for this target (this is just # truth points-1)}
 \item{"Std Dev" - This is the standard deviation of the truth data intervals (single number for all targets).}
 }
}
\description{
Table ennumerating gaps in truth data, measured in standard deviations
}
\examples{
summarize_truth_gaps(example1_scenario)
}
