% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_bearing_range_to_latlon.R
\name{transform_bearing_range_to_latlon}
\alias{transform_bearing_range_to_latlon}
\title{Transform bearing/range to lon/lat}
\usage{
transform_bearing_range_to_latlon(referenceData, relativeSensorData)
}
\arguments{
\item{referenceData}{data frame containing the longitude and latitude position of the sensor system.  MUST contain:
\itemize{
  \item{time: time of measurement (POSIX)}
  \item{lat: latitude at time of measurement (degrees)}
  \item{lon: longitude at time of measurement (degrees)}
  \item{alt: altitude at time of measurement (meters)}
  }}

\item{relativeSensorData}{data frame containing sensor readings of targets.  MUST contain:
\itemize{
  \item{time: time of measurement (POSIX)}
  \item{bearing: bearing to target (degrees azimuth)}
  \item{range: range to target (meters)}
  \item{alt: altitude of target (meters)}
  \item{trackNum: (factor) name of sensor track.  Number is recommended}
  }}
}
\value{
data frame with the lat and lon position of the targets.  Variables include:
\itemize{
  \item{time: time of measurement (POSIX)}
  \item{trackNum: (factor) name of sensor track.  Number is recommended}
  \item{lat: latitude (degrees)}
  \item{lon: longitude (degrees)}
  \item{alt: altitude (meters)}
 }
}
\description{
This function takes sensor data given in bearing/range/alt units and converts it to latitude, longitude, altitude based on the sensor system position.  Nautilus uses latitude and longitude, so units must be converted. The function will correct for sensor altitude in a purely additive manner (i.e., if sensor is at 1,000 feet, and tgt is 2,000 feet above, then tgt altitude is 3,000 feet).
}
