% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_offset_to_latlon.R
\name{transform_offset_to_latlon}
\alias{transform_offset_to_latlon}
\title{Transform offset to lon/lat}
\usage{
transform_offset_to_latlon(referenceData, relativeSensorData)
}
\arguments{
\item{referenceData}{data frame containing the longitude and latitude position of the sensor system.  MUST contain:
\itemize{
  \item{time: time of measurement (POSIX)}
  \item{lat: latitude at time of measurement (degrees)}
  \item{lon: longitude at time of measurement (degrees)}
  \item{alt: altitude at time of measurement (meters)}
  }}

\item{relativeSensorData}{data frame containing sensor readings of targets.  MUST contain:
\itemize{
  \item{time: time of measurement (POSIX)}
  \item{NorthOffset: meters North of sensor}
  \item{EastOffset: meters East of sensor}
  \item{UpOffset: meters above sensor}
  \item{trackNum: (factor) name of sensor track.  Number is recommended}
  }}
}
\value{
data frame with the lat and lon position of the targets.  Variables include:
\itemize{
  \item{time: time of measurement (POSIX)}
  \item{trackNum: (factor) name of sensor track.  Number is recommended}
  \item{lat: latitude (degrees)}
  \item{lon: longitude (degrees)}
  \item{alt: altitude (meters)}
  \item{any additional columns included in relativeSensorData (minus North/East/Up offset)}
 }
}
\description{
This function takes sensor data given in offset units (meters north / meters east of sensor / meters up from sensor location) and converts it to latitude and longitude based on the sensor system position.  Nautilus uses latitude and longitude, so units must be converted.
}
