% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_input_data.R
\name{verify_input_data}
\alias{verify_input_data}
\title{Checks the input format for create_scenario}
\usage{
verify_input_data(myData, columnNames, factorNames, stopOnFail = TRUE)
}
\arguments{
\item{myData}{object to be checked for correct format}

\item{columnNames}{column names that must be in myData}

\item{factorNames}{columns in myData that must be factors}

\item{stopOnFail}{Default = TRUE  When TRUE, function STOPS (bombs out) when dataframe fails a test and prints message about why. When FALSE, function simply returns FALSE.}
}
\value{
Bolean.  TRUE if input is in correct format, FALSE if input is in incorrect form and stopOnFail=FALSE. If stopOnFail=TRUE, then this will bomb out and return nothing
}
\description{
This function is called by create_scenario to check input dataframes to make sure they're compatible with scenarioMaker.

 It checks for the following:

* myData is a data frame

* myData is not empty

* myData has all of the columns in columnNames

* the columns named in factorName are factors
}
