% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyz_to_lla.R
\name{xyz_to_lla}
\alias{xyz_to_lla}
\title{Convert cartesian coordinates to Earth lon/lat/alt}
\usage{
xyz_to_lla(x, y, z)
}
\arguments{
\item{x}{cartesian x}

\item{y}{cartesian y}

\item{z}{cartesian z}
}
\value{
A list of 3 numbers where the 3 entries correspond to lon (degrees), lat (degrees), and altitude (meters) on Earth
}
\description{
This function converts cartesian coordinates to Earth lat/lon/alt (WGS84).  Cartesian coordinates are from the center of Earth.  Meant to be used in conjunction with \code{\link{lla_to_xyz}}
}
\examples{
scenarioMaker:::xyz_to_lla(12000, 24000, 36000)

}
\keyword{internal}
