% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_named_colors.R
\name{get_named_colors}
\alias{get_named_colors}
\title{Get named colors for plotting (plotly)}
\usage{
get_named_colors(scenario, idList, defaultColorList, useDefaultColors = FALSE)
}
\arguments{
\item{scenario}{must contain either or both: targetTruth, ownShipTruth}

\item{idList}{list of truthIDs or trackNums}

\item{defaultColorList}{list of default colors - must be as long as idList and in the same order (if you care)!}
}
\value{
plotly object
}
\description{
Internal function used by plotly functions.  Extracts the colors that should be used in the plot from platformInfo (contained in the scenario).  Alternatively, you can set useDefaultcolors = TRUE.
'
}
