% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_truth_data.R
\name{plot_truth_data}
\alias{plot_truth_data}
\title{Plot truth data}
\usage{
plot_truth_data(scenario, offset = 0.01, textSize = 3, hideLegend = FALSE,
  useDefaultColors = FALSE)
}
\arguments{
\item{scenario}{must contain either or both: targetTruth, ownShipTruth}

\item{offset}{amount to offset labels (in degrees latitude)}

\item{textSize}{size of label text}

\item{hideLegend}{hides the legend to make the graph easier to read (default=FALSE)}

\item{useDefaultColors}{(default=FALSE) if you DO have platformInfo filled in, but want to ignore the colors and use defaults, set this to true}
}
\value{
ggplot object
}
\description{
This figure shows an overview of the test and plots the ownship data (thin black line) and/or the truth data (solid lines). The target labels (likely characters or names) are placed above the the first point in time by an amount equal to the offset (in degrees latitude). The size of the label text is given by textSize.
}
\examples{
plot_truth_data(
 scenario=example1_scenario,
 offset = 0.01,
 textSize = 3,
 hideLegend = FALSE,
 useDefaultColors = FALSE)
}
